Update bindgen to version 0.69.4
The previous bindgen version (0.53.1) was vulnerable due to its
dependency shlex@^0.1 (all dependencies shlex@<1.3.0 are vulnerable
according to https://github.com/advisories/GHSA-r7qv-8r2h-pg27).

This commit updates the bindgen version to the last one published to
date that has shlex@^1 as a dependency, and so should not be vulnerable
anymore.